+++ native/native.c
@@ -26,5 +26,6 @@
 #include "base64.c"
 #include "map.c"
 #include "thrift.c"
+#include "thrift_binary.c"
 #include "thrift_compact.c"
 #include "thrift_skip.c"
