+++ native/xthrift.h
@@ -17,56 +17,56 @@
 #include "native.h"
 #include "map.h"
 
-#ifndef THRIFT_H
-#define THRIFT_H
+#ifndef XTHRIFT_H
+#define XTHRIFT_H
 
-#define F_ALLOW_UNKNOWN     1ull
-#define F_WRITE_DEFAULT     (1ull << 1)
-#define F_ENABLE_VM         (1ull << 2)
-#define F_ENABLE_HM         (1ull << 3)
-#define F_ENABLE_I2S        (1ull << 4)
-#define F_WRITE_REQUIRE     (1ull << 5)
-#define F_NO_BASE64         (1ull << 6)
-#define F_WRITE_OPTIONAL    (1ull << 7)
-#define F_TRACE_BACK        (1ull << 8)
+#define F_ALLOW_UNKNOWN 1ull
+#define F_WRITE_DEFAULT (1ull << 1)
+#define F_ENABLE_VM (1ull << 2)
+#define F_ENABLE_HM (1ull << 3)
+#define F_ENABLE_I2S (1ull << 4)
+#define F_WRITE_REQUIRE (1ull << 5)
+#define F_NO_BASE64 (1ull << 6)
+#define F_WRITE_OPTIONAL (1ull << 7)
+#define F_TRACE_BACK (1ull << 8)
 
 #define THRIFT_VERSION_MASK 0xffff0000ul
-#define THRIFT_VERSION_1    0x80010000ul
+#define THRIFT_VERSION_1 0x80010000ul
 
 typedef int32_t tmsg;
 
-#define TMESSAGE_TYPE_INVALID   0
-#define TMESSAGE_TYPE_CALL      1
-#define TMESSAGE_TYPE_REPLY     2
+#define TMESSAGE_TYPE_INVALID 0
+#define TMESSAGE_TYPE_CALL 1
+#define TMESSAGE_TYPE_REPLY 2
 #define TMESSAGE_TYPE_EXCEPTION 3
-#define TMESSAGE_TYPE_ONEWAY    4
+#define TMESSAGE_TYPE_ONEWAY 4
 
 typedef uint8_t ttype;
 
-#define TTYPE_EINVAL    (ttype)0xff
-#define TTYPE_STOP      (ttype)0
-#define TTYPE_VOID      (ttype)1
-#define TTYPE_BOOL      (ttype)2
-#define TTYPE_BYTE      (ttype)3
-#define TTYPE_I08       (ttype)3
-#define TTYPE_DOUBLE    (ttype)4
-#define TTYPE_I16       (ttype)6
-#define TTYPE_I32       (ttype)8
-#define TTYPE_I64       (ttype)10
-#define TTYPE_STRING    (ttype)11
-#define TTYPE_UTF7      (ttype)11
-#define TTYPE_STRUCT    (ttype)12
-#define TTYPE_MAP       (ttype)13
-#define TTYPE_SET       (ttype)14
-#define TTYPE_LIST      (ttype)15
-#define TTYPE_UTF8      (ttype)16
-#define TTYPE_UTF16     (ttype)17
-#define TTYPE_LAST      TTYPE_UTF16+1
+#define TTYPE_EINVAL (ttype)0xff
+#define TTYPE_STOP (ttype)0
+#define TTYPE_VOID (ttype)1
+#define TTYPE_BOOL (ttype)2
+#define TTYPE_BYTE (ttype)3
+#define TTYPE_I08 (ttype)3
+#define TTYPE_DOUBLE (ttype)4
+#define TTYPE_I16 (ttype)6
+#define TTYPE_I32 (ttype)8
+#define TTYPE_I64 (ttype)10
+#define TTYPE_STRING (ttype)11
+#define TTYPE_UTF7 (ttype)11
+#define TTYPE_STRUCT (ttype)12
+#define TTYPE_MAP (ttype)13
+#define TTYPE_SET (ttype)14
+#define TTYPE_LIST (ttype)15
+#define TTYPE_UTF8 (ttype)16
+#define TTYPE_UTF16 (ttype)17
+#define TTYPE_LAST TTYPE_UTF16+1
 
 typedef uint16_t vm_em;
-#define VM_NONE         0
-#define VM_JSCONV       101 // NOTICE: must be same with internal/native/types.VM_JSCONV
-#define VM_INLINE_MAX   255
+#define VM_NONE 0
+#define VM_JSCONV 101 // NOTICE: must be same with internal/native/types.VM_JSCONV
+#define VM_INLINE_MAX 255
 
 typedef struct tTypeDesc tTypeDesc;
 
@@ -220,6 +220,10 @@ typedef struct
 } J2TStateMachine;
 
 #define SIZE_J2TEXTRA sizeof(J2TExtra)
+
+
+uint64_t buf_malloc(_GoSlice *buf, size_t size);
+uint64_t j2t_fsm_exec(J2TStateMachine *self, _GoSlice *buf, const _GoString *src, uint64_t flag);
 
 #define J2T_VAL 0
 #define J2T_ARR 1
@@ -387,6 +391,8 @@ typedef struct
     xprintf("[DEBUG_PRINT_STATE_EX_%d] STATE{sp:%d, st:%d, jp:%d, td:%s, ex:[%d,%d,%d]} POS:%d CHAR:%c BUF{\n\tbuf:%l,\n\tlen:%d, cap:%d}\n", i, sp, vt->st, vt->jp, vt->td == NULL ? &(_GoString){} : &vt->td->name, *(uint64_t *)(&vt->ex), *((uint64_t *)(&vt->ex) + 1), *((uint64_t *)(&vt->ex) + 2), p, ch, buf, buf->len, buf->cap);
 
 
+
+
 typedef struct {
     void *resv0;
     void *resv1;
@@ -394,69 +400,6 @@ typedef struct {
     void *resv3;
 } vt_base;
 
-#define VT_J2TSM(vtb)       (vtb.resv0)
-#define VT_OUTBUF(vtb)      (vtb.resv1)
-#define VT_TC_STATE(vtb)    (vtb.resv3)
-#define VT_TB_STATE(vtb)    (vtb.resv3)
-
-#define TIMPL_MENC_EXTRA(tproto, tname) \
-    typedef struct { \
-        uint64_t    (*write_default_or_empty)(tproto, const tFieldDesc *, long); \
-        uint64_t    (*write_data_count)(tproto, size_t n); \
-        size_t      (*write_data_count_max_length)(tproto); \
-    } tname
-
-#define TIMPL_MENC_BASE(tproto, tname) \
-    typedef struct { \
-        uint64_t (*write_message_begin)(tproto, const _GoString, int32_t, int32_t); \
-        uint64_t (*write_message_end)(tproto); \
-        uint64_t (*write_struct_begin)(tproto); \
-        uint64_t (*write_struct_end)(tproto); \
-        uint64_t (*write_field_begin)(tproto, ttype, int16_t); \
-        uint64_t (*write_field_end)(tproto); \
-        uint64_t (*write_field_stop)(tproto); \
-        \
-        uint64_t (*write_map_n)(tproto, ttype, ttype, size_t); \
-        uint64_t (*write_map_begin)(tproto, ttype, ttype, size_t *); \
-        uint64_t (*write_map_end)(tproto, size_t, size_t); \
-        \
-        uint64_t (*write_list_n)(tproto, ttype, size_t); \
-        uint64_t (*write_list_begin)(tproto, ttype, size_t *); \
-        uint64_t (*write_list_end)(tproto, size_t, size_t); \
-        \
-        uint64_t (*write_set_n)(tproto, ttype, size_t); \
-        uint64_t (*write_set_begin)(tproto, ttype, size_t *); \
-        uint64_t (*write_set_end)(tproto, size_t, size_t); \
-        \
-        uint64_t (*write_bool)(tproto, bool); \
-        uint64_t (*write_byte)(tproto, char); \
-        uint64_t (*write_i16)(tproto, int16_t); \
-        uint64_t (*write_i32)(tproto, int32_t); \
-        uint64_t (*write_i64)(tproto, int64_t); \
-        uint64_t (*write_double)(tproto, double); \
-        uint64_t (*write_string)(tproto, const char *, size_t); \
-        uint64_t (*write_binary)(tproto, const _GoSlice); \
-    } tname
-
-#define TIMPL_MENC(tproto, tname) \
-    TIMPL_MENC_EXTRA(tproto, tname##_extra); \
-    TIMPL_MENC_BASE(tproto, tname);
 
 
-#define TIMPL_IENC(tproto, iname, tname) \
-    /* Interface implementation of (iname) */ \
-    \
-    TIMPL_MENC(tproto, tname); \
-    typedef struct { \
-        vt_base base; \
-        tname method; \
-        tname##_extra extra; \
-    } iname
-
-TIMPL_IENC(void *, vt_ienc, vt_menc);
-
-uint64_t buf_malloc(_GoSlice *buf, size_t size);
-uint64_t bm_malloc_reqs(_GoSlice *cache, ReqBitMap src, ReqBitMap *copy, long p);
-uint64_t j2t2_fsm_exec(vt_ienc *self, const _GoString *src, uint64_t flag);
-
-#endif // THRIFT_H
+#endif // XTHRIFT_H
